make rules to pull/push from overleaf

+++ workflow/utils/overleaf.py
@@ -5,6 +5,7 @@ import os
 import shutil
 import snakemake
 from pathlib import Path
+from urllib.parse import quote
 
 
 def get_overleaf_credentials(
@@ -21,28 +22,18 @@ def get_overleaf_credentials(
             # TODO
             raise exceptions.MissingOverleafCredentials()
         else:
-            # TODO: Replace other symbols in user's password? Overleaf recommends
-            # not using special symbols in the password, which is pretty silly.
-            # I bet we can escape the problematic ones here.
-            # See https://www.overleaf.com/learn/how-to/Troubleshooting_git_bridge_problems
-            val = val.replace("@", "%40")
+            # Replace special characters in the credentials
+            val = quote(val, safe="")
             creds.append(val)
 
     return creds
 
 
-def setup():
+def clone(project_id):
 
     # Logging
     logger = logging.get_logger()
-
-    # Get the Overleaf project id
-    project_id = snakemake.workflow.config["overleaf"].get("id", None)
-    if not project_id:
-        # No Overleaf integration for this project
-        return
-    else:
-        logger.info("Fetching Overleaf repo...")
+    logger.info("Cloning Overleaf repo...")
 
     # Set up a fresh temp directory
     if paths.overleaf.exists():
@@ -53,7 +44,7 @@ def setup():
     overleaf_email, overleaf_password = get_overleaf_credentials()
     url = f"https://{overleaf_email}:{overleaf_password}@git.overleaf.com/{project_id}"
 
-    # Set up a local version of the repo. We don't clone it to avoid
+    # Set up a local version of the repo. We don't actually _clone_ it to avoid
     # storing the url containing the password in .git/config
     run(["git", "init"], cwd=str(paths.overleaf))
 
@@ -67,20 +58,28 @@ def setup():
 
 def push_files(files, project_id):
 
-    # Logging
+    # Setup logging
     logger = logging.get_logger()
 
-    # Repo info
-    overleaf_email, overleaf_password = get_overleaf_credentials()
-    url = f"https://{overleaf_email}:{overleaf_password}@git.overleaf.com/{project_id}"
+    # Clone the repo
+    clone(project_id)
 
     # Process each file
+    skip = []
     for file in files:
 
         # Copy it to the local version of the repo
+        if not Path(file).exists():
+            skip.append(file)
+            continue
         file = Path(file).resolve()
         remote_file = paths.overleaf / file.relative_to(paths.tex)
-        shutil.copy(file, remote_file)
+        if file.is_dir():
+            if remote_file.exists():
+                shutil.rmtree(remote_file)
+            shutil.copytree(file, remote_file)
+        else:
+            shutil.copy(file, remote_file)
 
         # git-add the file
         run(
@@ -88,11 +87,17 @@ def push_files(files, project_id):
             cwd=str(paths.overleaf),
         )
 
+    # Remove missing files from the list
+    if skip:
+        skip_list = " ".join([str(s) for s in skip])
+        logger.warn(f"Skipping missing file(s): {skip_list}")
+        files = list(set(files) - set(skip))
+
     # Commit callback
     def callback(code, stdout, stderr):
         if stdout:
             logger.debug(stdout)
-        file_list = " ".join(files)
+        file_list = " ".join([str(s) for s in files])
         if code != 0:
             if "nothing to commit" in stdout:
                 logger.info(f"No changes to commit to Overleaf: {file_list}")
@@ -119,8 +124,35 @@ def push_files(files, project_id):
     )
 
     # Push (again being careful about secrets)
+    overleaf_email, overleaf_password = get_overleaf_credentials()
+    url = f"https://{overleaf_email}:{overleaf_password}@git.overleaf.com/{project_id}"
     run(
         ["git", "push", url, "master"],
         cwd=str(paths.overleaf),
         secrets=[overleaf_email, overleaf_password],
     )
+
+
+def pull_files(files, project_id):
+
+    # Clone the repo
+    clone(project_id)
+
+    # Copy over the files
+    file_list = " ".join(files)
+    logger = logging.get_logger()
+    logger.info(f"Pulling changes from Overleaf: {file_list}")
+    logger.warn("Run `git status` to see what changed. Don't forget to commit!")
+    for file in files:
+        file = Path(file).absolute()
+        remote_file = (paths.overleaf / file.relative_to(paths.tex)).resolve()
+        if not remote_file.exists():
+            raise exceptions.MissingOverleafFile(
+                remote_file.relative_to(paths.overleaf)
+            )
+        if remote_file.is_dir():
+            if file.exists():
+                shutil.rmtree(file)
+            shutil.copytree(remote_file, file)
+        else:
+            shutil.copy(remote_file, file)
